+++ ceammc/ext/src/flow/flow_multiplex.h
@@ -2,11 +2,25 @@
 #define FLOW_MULTIPLEX_H
 
 #include "ceammc_object.h"
-
+#include "ceammc_proxy.h"
 using namespace ceammc;
 
 class FlowMultiplex : public BaseObject {
-    SizeTProperty* index_;
+public:
+    struct ControlInlet {
+        FlowMultiplex* pimpl;
+        void on_float(t_float f) { pimpl->proxy_float(f); }
+    };
+
+    using Inlet = InletProxy<FlowMultiplex>;
+    using ControlProxy = InletProxy<ControlInlet>;
+
+private:
+    IntProperty* index_;
+    const size_t total_;
+    std::vector<Inlet> inlets_;
+    ControlInlet ctrl_;
+    ControlProxy control_;
 
 public:
     FlowMultiplex(const PdArgs& args);
@@ -18,6 +32,14 @@ public:
     void onAny(t_symbol* sel, const AtomListView& args) override;
     void onData(const Atom& data) override;
     void onInlet(size_t idx, const AtomList& l) override;
+
+    const char* annotateInlet(size_t n) const override;
+
+    void proxy_any(Inlet* x, t_symbol* s, const AtomListView& v);
+    void proxy_float(t_float f);
+
+public:
+    static void initXletInfo();
 };
 
 void setup_flow_multiplex();